+++ src/presenters/pages/index.jsx
@@ -8,15 +8,46 @@ import {CurrentUserProvider, CurrentUserConsumer} from '../current-user.jsx';
 
 import Categories from '../categories.jsx';
 import Featured from '../featured.jsx';
+import OverlayVideo from '../overlays/overlay-video.jsx';
 import Questions from '../questions.jsx';
 import RandomCategories from '../random-categories.jsx';
 import RecentProjects from '../recent-projects.jsx';
-import WhatIsGlitch from '../what-is-glitch.jsx';
+
+const play = "https://cdn.hyperdev.com/6ce807b5-7214-49d7-aadd-f11803bc35fd%2Fplay.svg";
+const whatsGlitchWide = "https://cdn.glitch.com/f7224274-1330-4022-a8f2-8ae09dbd68a8%2Fwhats-glitch-wide.svg?1499885209761";
+const whatsGlitchNarrow = "https://cdn.glitch.com/f7224274-1330-4022-a8f2-8ae09dbd68a8%2Fwhats-glitch-narrow.svg?1499884900667";
+const free = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Ffree.svg?1499350845981";
+
+const whatsGlitchAlt = "Create a node app, or remix one. It updates as you type. Code with Friends!";
+  
+const WhatIsGlitch = () => (
+  <section className="what-is-glitch">
+    <h2>How It Works</h2>
+    <span>
+      <a href="https://glitch.com/about">
+        <figure title="How Glitch works">
+          <img className="wide" src={whatsGlitchWide} alt={whatsGlitchAlt}/>
+          <img className="narrow" src={whatsGlitchNarrow} alt={whatsGlitchAlt}/>
+        </figure>
+      </a>
+      <div>
+        And it's <img className="free" src={free} alt="free"/>.{' '}
+        <OverlayVideo>
+          <div className="button video">
+            <img className="play-button" src={play} alt="play"/>
+            <span>How it works in 1 minute</span>
+          </div>
+        </OverlayVideo>
+      </div>
+    </span>
+  </section>
+);
 
 const FOUNDED = 2000;
 const current = new Date().getFullYear();
 const age = current - FOUNDED;
 const logo = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Ffogcreek.svg";
+
 const ByFogCreek = () => (
   <section className="by-fogcreek" role="complementary">
     <h2>Made By Fog Creek</h2>
@@ -31,12 +62,10 @@ const ByFogCreek = () => (
 
 const MadeInGlitch = () => (
   <section className="made-in-glitch">
-    <p>
-      Of course, this site was made on Glitch too{' '}
-      <a href={getEditorUrl('community')} className="button button-link has-emoji">
-        View Source <span className="emoji carp_streamer"></span>
-      </a>
-    </p>
+    <p>Of course, this site was made on Glitch too</p>
+    <a href={getEditorUrl('community')} className="button button-link has-emoji">
+      View Source <span className="emoji carp_streamer"></span>
+    </a>
   </section>
 );
 
@@ -62,8 +91,8 @@ const IndexPageContainer = ({userModel, ...props}) => (
     <CurrentUserConsumer>
       {user => <IndexPage user={user} {...props}/>}
     </CurrentUserConsumer>
-    <script src="//fast.wistia.com/embed/medias/vskja9agqj.jsonp" async/>
-    <script src="//fast.wistia.com/assets/external/E-v1.js" async/>
+    <script src="//fast.wistia.com/embed/medias/vskja9agqj.jsonp"/>
+    <script src="//fast.wistia.com/assets/external/E-v1.js"/>
   </CurrentUserProvider>
 );
